+++ src/commands/make/Mail.ts
@@ -1,23 +1,22 @@
-import Common from '../../Common'
+import Common from '../../Common';
 
 export default class MakeMail extends Common {
-
   public static async run() {
-
-    let mailName = await this.getInput('Mail Name')
-    if (mailName.length == 0) {
-      this.showError('A mail name is required')
-      return
+    let mailName = await this.getInput('Mail Name');
+    if (mailName.length === 0) {
+      this.showError('A mail name is required');
+      return;
     }
 
-    let command = `make:mail ${mailName}`
+    const markdown = await this.getNoYes('Should this use markdown?');
+    let command = `make:mail ${mailName} ${markdown ? '--markdown' : ''}`;
 
-    this.execCmd(command, async (info) => {
+    this.execCmd(command, async info => {
       if (info.err) {
-        this.showError('Could not create the mailler', info.err)
+        this.showError('Could not create the mailer', info.err);
       } else {
-        await this.openFile(info.artisan.dir, '/app/Mail/' + mailName + '.php')
+        await this.openFile(info.artisan.dir, '/app/Mail/' + mailName + '.php');
       }
-    })
+    });
   }
 }