+++ src/commands/make/Migration.ts
@@ -1,39 +1,40 @@
-import Common from '../../Common'
+import Common from '../../Common';
 
 export default class MakeMigration extends Common {
-
   public static async run() {
     // Get the name of the controller to create
-    let migrationName = await this.getInput('Migration Name')
-    if (migrationName.length == 0) {
-      this.showError('A migration name is required')
-      return
+    let migrationName = await this.getInput('Migration Name');
+    if (migrationName.length === 0) {
+      this.showError('A migration name is required');
+      return;
     }
 
-    let createTable = false
-    let modifyTable = false
-    let tableName = ''
+    let createTable = false;
+    let modifyTable = false;
+    let tableName = '';
 
     // Determine if this is a resource controller or not
-    createTable = await this.getYesNo('Will this migration create a table?')
+    createTable = await this.getYesNo('Will this migration create a table?');
     if (!createTable) {
-      modifyTable = await this.getYesNo('Will this migration modify an existing table?')
+      modifyTable = await this.getYesNo('Will this migration modify an existing table?');
     }
 
     if (createTable || modifyTable) {
-      tableName = await this.getInput('What is the name of the table?')
+      tableName = await this.getInput('What is the name of the table?');
     }
 
-    let command = `make:migration "${migrationName}" ${createTable ? '--create=' + tableName : ''} ${modifyTable ? '--table=' + tableName : ''}`
+    let command = `make:migration "${migrationName}" ${createTable ? '--create=' + tableName : ''} ${
+      modifyTable ? '--table=' + tableName : ''
+    }`;
 
     // Generate the controller
-    this.execCmd(command, async (info) => {
+    this.execCmd(command, async info => {
       if (info.err) {
-        this.showError('Could not create the migration', info.err)
+        this.showError('Could not create the migration', info.err);
       } else {
-        let file = info.stdout.replace(/^.+:/ig, '').trim()
-        await this.openFile(info.artisan.dir, '/database/migrations/' + file + '.php')
+        let file = info.stdout.replace(/^.+:/gi, '').trim();
+        await this.openFile(info.artisan.dir, '/database/migrations/' + file + '.php');
       }
-    })
+    });
   }
 }