+++ src/commands/make/Notification.ts
@@ -1,23 +1,21 @@
-import Common from '../../Common'
+import Common from '../../Common';
 
 export default class MakeNotification extends Common {
-
   public static async run() {
-
-    let noteName = await this.getInput('Notification Name')
-    if (noteName.length == 0) {
-      this.showError('A notification name is required')
-      return
+    let noteName = await this.getInput('Notification Name');
+    if (noteName.length === 0) {
+      this.showError('A notification name is required');
+      return;
     }
 
-    let command = `make:notification ${noteName}`
+    let command = `make:notification ${noteName}`;
 
-    this.execCmd(command, async (info) => {
+    this.execCmd(command, async info => {
       if (info.err) {
-        this.showError('Could not create the notification', info.err)
+        this.showError('Could not create the notification', info.err);
       } else {
-        await this.openFile(info.artisan.dir, '/app/Notifications/' + noteName + '.php')
+        await this.openFile(info.artisan.dir, '/app/Notifications/' + noteName + '.php');
       }
-    })
+    });
   }
 }