+++ src/commands/make/Provider.ts
@@ -1,23 +1,21 @@
-import Common from '../../Common'
+import Common from '../../Common';
 
 export default class MakeProvider extends Common {
-
   public static async run() {
-
-    let providerName = await this.getInput('Provider Name')
-    if (providerName.length == 0) {
-      this.showError('A provider name is required')
-      return
+    let providerName = await this.getInput('Provider Name');
+    if (providerName.length === 0) {
+      this.showError('A provider name is required');
+      return;
     }
 
-    let command = `make:provider ${providerName}`
+    let command = `make:provider ${providerName}`;
 
-    this.execCmd(command, async (info) => {
+    this.execCmd(command, async info => {
       if (info.err) {
-        this.showError('Could not create the provider', info.err)
+        this.showError('Could not create the provider', info.err);
       } else {
-        await this.openFile(info.artisan.dir, '/app/Providers/' + providerName + '.php')
+        await this.openFile(info.artisan.dir, '/app/Providers/' + providerName + '.php');
       }
-    })
+    });
   }
 }