+++ src/commands/make/Request.ts
@@ -1,23 +1,21 @@
-import Common from '../../Common'
+import Common from '../../Common';
 
 export default class MakeRequest extends Common {
-
   public static async run() {
-
-    let requestName = await this.getInput('Request Name')
-    if (requestName.length == 0) {
-      this.showError('A request name is required')
-      return
+    let requestName = await this.getInput('Request Name');
+    if (requestName.length === 0) {
+      this.showError('A request name is required');
+      return;
     }
 
-    let command = `make:request ${requestName}`
+    let command = `make:request ${requestName}`;
 
-    this.execCmd(command, async (info) => {
+    this.execCmd(command, async info => {
       if (info.err) {
-        this.showError('Could not create the request', info.err)
+        this.showError('Could not create the request', info.err);
       } else {
-        await this.openFile(info.artisan.dir, '/app/Http/Requests/' + requestName + '.php')
+        await this.openFile(info.artisan.dir, '/app/Http/Requests/' + requestName + '.php');
       }
-    })
+    });
   }
 }